[XLA:GPU] Support fusion of dynamic-slice into triton gemm.
This change fuses dynamic-slice into triton gemm kernels provided the slice is taken along the major-most dimension, leaving all other dimensions the same. No further fusion occurs in operands of the dynamic slice, meaning the resulting triton gemm must take in all operands of the dynamic slice as parameters.

The original author of this CL is jvstokes and I modified it a bit.

PiperOrigin-RevId: 631438545